Store in seconds.
I'm not sure this is useful still - the full hydration test definitely
takes more than a second to run, but since we're just measuring what it
takes to produce those last messages it's less than a second. Maybe the
differential isn't a big deal in production.

+++ lib/dpul_collections/indexing_pipeline/index_metric.ex
@@ -5,7 +5,8 @@ defmodule DpulCollections.IndexingPipeline.IndexMetric do
   schema "index_metrics" do
     field :type, :string
     field :measurement_type, :string
-    field :duration, :decimal
+    # Duration in seconds
+    field :duration, :integer
 
     timestamps(type: :utc_datetime_usec)
   end
